requirements.txt
pygame>=2.5.2
mss>=9.0.1
PyAutoGUI>=0.9.54
Pillow>=10.1.0
opencv-python>=4.10.0.82
numpy>=1.26.0
playsound>=1.3.0
chess>=1.10.0
keyboard>=0.13.5
selenium>=4.9.0